Add Initials value to Contact
Hello, I noticed that the code for providing initials for the Contact avatar was in the `ContactListPage`, and that it gave the first two letters of `displayName` rather than initials. This PR is to add a method that computes a `Contact`'s initials within the model.

I'm still getting used to Dart as a language (as an iOS dev I use Swift), so if this isn't the best approach let me know. My first intuition was to extend `Contact`, however I'm not sure Dart supports that... Instead I tried to mimic a computed variable by creating this initials() function.

Would appreciate any feedback (particular in terms of syntax or creating extensions), and thanks for maintaining this framework!